requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
black==24.10.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.4
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
kombu==5.4.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2==2.9.10
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
